[#158836814] Add package 'ipa' for ipa data management. Add ipa name to es data
