+++ crawler/publiccode_model.go
@@ -2,7 +2,9 @@ package crawler
 
 // PublicCodeES describe the data in ElasticSearch that includes publiccode and meta informations.
 type PublicCodeES struct {
-	FileRawURL       string `json:"fileRawURL"`
+	FileRawURL      string `json:"fileRawURL"`
+	Amministrazione string `json:"amministrazione"`
+
 	Name             string `json:"name"`
 	ApplicationSuite string `json:"applicationSuite,omitempty"`
 	URL              string `json:"url"`
